factor connect.go from utils.go
